Free using `layout` for posix_memalign'ed blocks
Due to an issue in the Boehm collector [1], blocks allocated using
posix_memalign must pass the base pointer (regardless of alignment) to
GC_free in order to be manually freed.

We use the `layout` argument to determine whether a block was allocated
using posix_memalign, and if so, use Boehm's `GC_base` API to obtain the
base pointer.

[1]: https://github.com/ivmai/bdwgc/issues/589