Remove traces of Alloy from library/alloc
Having parts of GC code in both library/std and library/alloc is a
hangover from before and overly complicates linking. In order to apply
some of the optimisations that we want on for off-thread finalisation,
we must first fix this.

This commit simply shuffles code around and renames things.