Add a function for users to keep pointers alive
This is useful for advanced users, who may allow references to GC
objects to escape from their Gc<T> smart pointers. In such cases, one
must be careful to ensure that a finalizer is not run early because of
compiler optimisations (e.g. scalar replacement). This commit provides a
unified way to do this using the same mechanism as Gc::drop.