Refactor FSA error generation
This PR creates a single enum to represent the various kinds of FSA
errors that Rust programs may exhibit. This is a mostly refactoring
commit which hopes to simplify some more complex analysis that we want
to perform.

However, this does have one observable impact on the current state of
FSA in Alloy: we now emit different errors for types which are `!Send` +
`!Sync`, and those which are `!FinalizerSafe`.

On its own, this may seem extraneous, as it now means that types which
are `!Send + !Sync + !FinalizerSafe` will now generate two errors
instead of one. However, this vastly simplifies future changes to the
FSA algorithm, and makes it much easier to 'bail-out' of FSA early when
we find bad cases which would not maek sense to recurse further into
(e.g. standard library types).