Less granular FSA errors for standard library types
This does not weaken FSA -- the drop methods of standard library types
are still scanned line-by-line. However, if a standard library type
contains a drop method with an unsafe projection, a single generic error
for that type will be emitted (as opposed to errors on a per-line basis
as with user crates).