Use '#[rustc_fsa_entry_point]' attribute for FSA
Previously we hard-coded FSA to begin when it discovered calls to
`Gc::new`. However, with the `Gc::from` API we now have multiple `Gc`
constructors where FSA may need to start from. This new attribute allows
us to decorate the various constructors in `library/std/gc.rs` as
needed.

This also refactors some of the run MirPass code to make it a bit easier
to work with Rust lifetimes.