Add basic support for checking function calls in FSA
Previously, FSA couldn't follow function calls. So if FSA encountered a
drop method such as below:

```rust
fn drop(&mut self) {
    foo(&self);
}
```

It would always emit an error because it couldn't know whether `foo`
contained a projection into a non finalizer-safe field.

Now, FSA will try and resolve calls to their function definition, and if
the MIR is available, will follow them and check them for safety as with
drop methods.

Ideally, this would be a depth-first traversal, where any FSA errors
emitted follow (as closely as possible) the control flow of the program.
For example:

```rust
fn foo(x: T) {
    self.bad_projection // Error 1
    bar(&self)
    self.bad_projection // Error 3
}

fn bar(x: T) {
    self.bad_projection // Error 2

}

fn drop(&mut self) {
    foo(&self);
    self.bad_projection // Error 4
}
```

Should emit FSA errors in the following order:

    Error 1 (from `foo`)
    Error 2 (from `bar`)
    Error 3 (from `foo`)
    Error 4 (from `drop`)

Instead, we emit errors in a breadth-first traversal, where we enqueue
each call to be checked in the order we encounter them:

    Error 4 (from `drop`)
    Error 1 (from `foo`)
    Error 3 (from `foo`)
    Error 2 (from `bar`)

We do this because the MIR visiter (a rustc trait for walking a
function's MIR) is implemented using recursion, and if we're not
careful, and naively call nested visiters as we encounter calls we can
easily overflow the stack.

Short of re-implementing the MIR visitor to use iteration, we simply add
calls that we encounter to a queue to be processed immediately after
each visit.

As with the rest of FSA, this is conservative in its approach. We don't
(and obviously can't) support all types of function calls (e.g. from a
function pointer, or trait methods), but when such calls are encountered we
throw an error to be on the safe side.