Add compile-time option for logging Alloy stats
This can be used by our experiment runner to obtain relevant stats about
the totals:

* Finalizers registered
* Finalizers run
* GC objects allocated
* Boxed objects allocated
* Rc objects allocated
* Arc objects allocated
* GC cycles

This data is logged to a file when provided with the
`LOG_ALLOY_STATS=/path/to/file.log` environment variable.

Since this is a compile-time flag used when building Alloy, it has no
effect on the performance of Alloy programs when disabled.