Make `FinalizerSafe` an internal only trait
There were too many ways for users to disable FSA and it became too
confusing to be worthwhile to ensure the logic for all the ways that
they could interact is correct.

This removes the `FinalizerSafe` trait implementation from all types
except `&T` and `Gc<T>` types, where it will be used to track if a type
contains a reference.