Use `FinalizerSafe` to determine whether FSA can short-circuit
It's not safe to avoid checking `T::drop` if `T: Send + Sync`, this is
because `T` might contain a `Gc` field or a Rust reference which would
be unsafe to use inside its drop method.

We thus repurpose the `FinalizerSafe` trait to act as a way of
'poisoning' types which contain such references. With this in place,
we can know whether or not we need to traverse into the field types or
not depending on whether `T` implements `Send + Sync + FinalizerSafe`.