Emit FSA error for unions with manually drop fields
For reasons that are not quite clear to me, trying to monomorphize a
manually droppable union field causes an internal compile error in
rustc. I think for now the best way to deal with this is to throw an
error and require the user to manually mark such types as finalizer safe
or not depending on their use-case.

Ideally we could improve on this at a later date as there is no reason
in theory why we shouldn't be able to support union types.