Switch to using bitflags for FSA
There are now various combinations of possible checks that FSA can
perform which can change dynamically based on previously identified
errors (e.g. the discovery of a thread-local means that we no longer
check for other things). By passing this around as bit flags it becomes
easier to write out the conditional logic.